chore: delete old hook
